Break out SurveyMissionItem->ComplexMissionItem class hierarchy
