+++ src/MissionManager/MissionItem.h
@@ -27,7 +27,7 @@
 #include "QmlObjectListModel.h"
 #include "MissionCommands.h"
 
-class ComplexMissionItem;
+class SurveyMissionItem;
 class SimpleMissionItem;
 class MissionController;
 #ifdef UNITTEST_BUILD
@@ -128,7 +128,7 @@ private:
     static const char*  _jsonAutoContinueKey;
     static const char*  _jsonCoordinateKey;
 
-    friend class ComplexMissionItem;
+    friend class SurveyMissionItem;
     friend class SimpleMissionItem;
     friend class MissionController;
 #ifdef UNITTEST_BUILD
