+++ src/MissionManager/SimpleMissionItem.cc
@@ -124,7 +124,7 @@ SimpleMissionItem::SimpleMissionItem(const SimpleMissionItem& other, QObject* pa
 
 const SimpleMissionItem& SimpleMissionItem::operator=(const SimpleMissionItem& other)
 {
-    static_cast<VisualMissionItem&>(*this) = other;
+    VisualMissionItem::operator=(other);
 
     setRawEdit(other._rawEdit);
     setDirty(other._dirty);
